# PERF: Improve performance of Series.isin() on sets