--- a/pandas/core/dtypes/common.py
+++ b/pandas/core/dtypes/common.py
@@ -17,7 +17,7 @@ from pandas.core.dtypes.inference import (  # noqa:F401
     is_array_like, is_bool, is_complex, is_decimal, is_dict_like, is_file_like,
     is_float, is_hashable, is_integer, is_interval, is_iterator, is_list_like,
     is_named_tuple, is_nested_list_like, is_number, is_re, is_re_compilable,
-    is_scalar, is_sequence, is_string_like)
+    is_set_like, is_scalar, is_sequence, is_string_like)
 
 _POSSIBLY_CAST_DTYPES = {np.dtype(t).name
                          for t in ['O', 'int8', 'uint8', 'int16', 'uint16',