--- a/pandas/tests/dtypes/test_inference.py
+++ b/pandas/tests/dtypes/test_inference.py
@@ -24,8 +24,8 @@ from pandas.core.dtypes import inference
 from pandas.core.dtypes.common import (
     ensure_categorical, ensure_int32, is_bool, is_datetime64_any_dtype,
     is_datetime64_dtype, is_datetime64_ns_dtype, is_datetime64tz_dtype,
-    is_float, is_integer, is_number, is_scalar, is_scipy_sparse,
-    is_set_like, is_timedelta64_dtype, is_timedelta64_ns_dtype)
+    is_float, is_integer, is_number, is_scalar, is_scipy_sparse, is_set_like,
+    is_timedelta64_dtype, is_timedelta64_ns_dtype)
 
 import pandas as pd
 from pandas import (
@@ -105,7 +105,6 @@ def test_is_list_like_disallow_sets(maybe_list_like):
 
 @pytest.mark.parametrize('obj,expected', [
     ({1, 2}, True),
-    (set([1, 2]), True),
     (set(), True),
     (set, False),
     ([1, 2], False),